test(sidebar): move expanded session pins into their sibling home
